Update opportunities_benefits.rst
we usually make "homeowner" one word, but I find it as two words in various places in the guide.  Unless you have a good reason to keep it as two words, let's be consistent and make it one word.

Also "home performance" should not be capitalized generally.  Home Performance with ENERGY STAR should be capitalized.